add function to calculate a single expression
calculateNumbers() checks for a single math
expression in the mainDisplay.textContent, then
split its contents into an array of
[number, operator,number] then pass it into
operate() function to calculate the expression.

it also updates the mainDisplay.textContent with
the return value of operate().

calculateNumbers() only process one expression at
the time regardless of number of expressions on
mainDisplay.textContent.